requirements.txt
python-telegram-bot
requests
mysql.connector
cowin-vaccine-api
telegram
python-dotenv